requirements.txt
Python 3.13.7
pytest 8.4.2
pluggy 1.6.0
allure-pytest 2.15.3
requests 2.32.5
selenium 4.38.0